# product-service/docs/swagger.yaml
definitions:
...
  request.CreateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      icon:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
      parent_id:
        type: integer
      status:
        type: boolean
    required:
    - icon
    - name
    - status
    type: object
...
  request.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      icon:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
      parent_id:
        type: integer
      status:
        type: boolean
    required:
    - icon
    - name
    - status
    type: object
  response.CategoryListAdminResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        type: boolean
      total_product:
        type: integer
    type: object
  response.CategoryListHomeResponse:
    properties:
      icon:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  response.CategoryListShopChildResponse:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
  response.CategoryListShopResponse:
    properties:
      child:
        items:
          $ref: '#/definitions/response.CategoryListShopChildResponse'
        type: array
      name:
        type: string
      slug:
        type: string
    type: object
  response.CategoryResponse:
    properties:
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        type: boolean
    type: object
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  response.DefaultResponseWithPagination:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
      total_page:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /admin/categories:
    get:
      consumes:
      - application/json
      description: Get paginated list of categories for admin
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: 'Order by field (default: created_at)'
        in: query
        name: orderBy
        type: string
      - description: 'Order type ASC or DESC (default: desc)'
        in: query
        name: orderType
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponseWithPagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.CategoryListAdminResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all categories (admin)
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Create Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create category (admin)
      tags:
      - categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict - Category has products
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete category (admin)
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get category detail by ID for admin
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get category by ID (admin)
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update category (admin)
      tags:
      - categories
  /admin/categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get category detail by slug for admin
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get category by slug (admin)
      tags:
      - categories
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get all published categories with nested children for shop page
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.CategoryListShopResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get all categories with children (public)
      tags:
      - categories
  /categories/featured:
    get:
      consumes:
      - application/json
      description: Get all published parent categories for home page
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.CategoryListHomeResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get featured categories (public)
      tags:
      - categories
...
swagger: "2.0"